Move _PublisherThread to BlockPublisher
