+++ validator/sawtooth_validator/state/batch_tracker.py
@@ -18,7 +18,7 @@ from threading import RLock
 from sawtooth_validator.journal.timed_cache import TimedCache
 from sawtooth_validator.journal.chain import ChainObserver
 from sawtooth_validator.execution.executor import InvalidTransactionObserver
-from sawtooth_validator.journal.journal import PendingBatchObserver
+from sawtooth_validator.journal.publisher import PendingBatchObserver
 from sawtooth_validator.protobuf.client_pb2 import ClientBatchStatus
